Fix non-interactive compilation
`.znap.compile` checks the `funcstack`, with the intent of determining
whether or not to output user visible messages -- they should only be
output when called directly from the `znap` CLI tool.

Unfortunately, this check was written in a way that means that when
called from outside of the main znap function (eg, compdump background
compilation, auto-compile), no compilation happened.

This change re-orders the checks so that `.znap.compile` works when
called from other parts of znap.

Fixes #267